+++ Runtime/Menus/MainMenu/AbstractMainMenuOptionsScreen.cs
@@ -6,7 +6,6 @@ namespace OneM.UISystem
     {
         [Header("Buttons")]
         [SerializeField] protected ActionButton continueButton;
-        [SerializeField] protected ActionButton startButton;
         [SerializeField] protected ActionButton loadButton;
         [SerializeField] protected ActionButton optionsButton;
         [SerializeField] protected ActionButton exitButton;
@@ -20,7 +19,6 @@ namespace OneM.UISystem
             base.SubscribeEvents();
 
             continueButton.OnClicked += HandleContinueButtonClicked;
-            startButton.OnClicked += HandleStartButtonClicked;
             loadButton.OnClicked += HandleLoadButtonClicked;
             optionsButton.OnClicked += HandleSettingsButtonClicked;
             exitButton.OnClicked += HandleExitButtonClicked;
@@ -31,7 +29,6 @@ namespace OneM.UISystem
             base.UnsubscribeEvents();
 
             continueButton.OnClicked -= HandleContinueButtonClicked;
-            startButton.OnClicked -= HandleStartButtonClicked;
             loadButton.OnClicked -= HandleLoadButtonClicked;
             optionsButton.OnClicked -= HandleSettingsButtonClicked;
             exitButton.OnClicked -= HandleExitButtonClicked;
@@ -41,12 +38,11 @@ namespace OneM.UISystem
         {
             var isContinueAvailable = await IsContinueAvailableAsync();
             continueButton.gameObject.SetActive(isContinueAvailable);
-            firstInput = isContinueAvailable ? continueButton.gameObject : startButton.gameObject;
+            firstInput = isContinueAvailable ? continueButton.gameObject : loadButton.gameObject;
         }
 
         protected abstract Awaitable<bool> IsContinueAvailableAsync();
         protected abstract void HandleContinueButtonClicked();
-        protected abstract void HandleStartButtonClicked();
 
         protected virtual void HandleLoadButtonClicked() => OpenCloseableScreen(loadScreen);
         protected virtual void HandleSettingsButtonClicked() => OpenCloseableScreen(optionsScreen);